requirements.txt
google-genai>=0.5.0
ddgs>=0.1.0
flask